requirements.txt
requests==2.32.5
pyTelegramBotAPI==4.28.0
python-dotenv==1.1.1